Require python 3.10.0
This might be a bit too strict, but I know that at least 3.6 is
required since I format strings using the new string literals
syntax that came in this release.

But 3.10.0 is the most recent version, and the one I used to develop
the script. Let's play safe and go with that.