Replace moves with modifications if the move has already been applied
When upgrading from 4.27.2 to 5.0.0-preview-1 git reports two files as
moved when they have in fact already been moved in plastic.

These two files
LinuxCrossToolchain.arm-unknown-linux-gnueabihf.cmake
LinuxCrossToolchain.x86_64-unknown-linux-gnu.cmake

Are reported from beeing moved from linux to Linux subfolders
of Engine/Source/ThirdParty/PhysX3/Externals/CMakeModules.

However in the plastic repo they aready exist in the Linux folder.

The files have lived in that folder ever since we copied the first UE
release into plastic, 4.27.1. And the same Linux folder exist in the zip file
you download for that release.

This must be due to some inconcistencies between the git repo, and the
zipped file releases.

The graceful way of handling this is to simply not move the files, but
instead copy their contents for the new release, in case it has changed.